Use an InsertMultiple trait.
