+++ common/persistence/sql/dbal/class.Driver.php
@@ -27,6 +27,7 @@
  */
 class common_persistence_sql_dbal_Driver implements common_persistence_sql_Driver
 {
+    use common_persistence_sql_InsertMultiple;
 
     /**
      * @var \Doctrine\DBAL\Connection
@@ -126,17 +127,6 @@ class common_persistence_sql_dbal_Driver implements common_persistence_sql_Drive
             $cleanColumns[$this->getPlatForm()->quoteIdentifier($columnName)] = $value;
         }
         return $this->connection->insert($tableName, $cleanColumns);
-    }
-    
-    public function insertMultiple($tableName, array $data)
-    {
-        $returnValue = 0;
-        
-        foreach ($data as $d) {
-            $returnValue += intval($this->insert($tableName, $d));
-        }
-        
-        return $returnValue;
     }
     
     /**
